Change bindings to work with any fuzzy-finder
Selecta uses raw ASCII escape sequences to run while vim is still
running. Fzf, by contrast, uses ncurses and freezes the vim display if
we attempt to simply run it with `:system`.

Now we create a tempfile, run the command using `:execute` and pipe the
output to our tempfile. We then read the contents and edit the resulting
file.